Quiz and Question Service is Developed

+++ src/main/java/com/codemind/QuizApp/service/QuestionService.java
@@ -3,43 +3,69 @@ package com.codemind.QuizApp.service;
 import com.codemind.QuizApp.dao.QuestionDAO;
 import com.codemind.QuizApp.entity.Question;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class QuestionService {
 
+    @Autowired
     private QuestionDAO questionDAO;
 
-    @Autowired
-    public QuestionService(QuestionDAO questionDAO) {
-        this.questionDAO = questionDAO;
+
+
+    public ResponseEntity<List<Question>> getAllQuestions() {
+        try {
+            return new ResponseEntity<>(questionDAO.findAll(), HttpStatus.OK);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+        return new ResponseEntity<>(new ArrayList<>(),HttpStatus.BAD_REQUEST);
     }
 
-    public List<Question> getAllQuestions() {
-        return questionDAO.findAll();
-    }
-
-    public List<Question> getQuestionByCategory(String category){
-        return questionDAO.findByCategory(category);
-    }
-
-    public String addQuestion(Question question) {
-        questionDAO.save(question);
-        return "New ID which is generated : "+question.getId();
-    }
-
-    public String updateQuestion(Question question) {
-
-        questionDAO.save(question);
-
-        return "Update data which Id is : "+question.getId()+" and Question is : "+question.getQuestionTitle();
+    public ResponseEntity<List<Question>> getQuestionByCategory(String category){
+        try {
+            return new ResponseEntity<>(questionDAO.findByCategory(category),HttpStatus.OK);
+        }catch (Exception e){
+            e.printStackTrace();
+        }
+        return new ResponseEntity<>(new ArrayList<>(),HttpStatus.BAD_REQUEST);
 
     }
 
-    public String deleteQuestion(int id) {
-        questionDAO.deleteById(id);
-        return "Question is deleted which ID is : "+id;
+    public ResponseEntity<String> addQuestion(Question question) {
+        if(question.getQuestionTitle() != null){
+            questionDAO.save(question);
+            return new ResponseEntity<>("New ID which is generated : "+question.getId(),HttpStatus.CREATED);
+        }
+        return new ResponseEntity<>("Data is Null",HttpStatus.BAD_REQUEST);
+
+    }
+
+    public ResponseEntity<String> updateQuestion(Question question) {
+
+             try{
+                questionDAO.save(question);
+
+                return new ResponseEntity<>("Update data which Id is : " + question.getId() + " and Question is : " + question.getQuestionTitle(), HttpStatus.CREATED);
+            }catch (Exception e){
+                 e.printStackTrace();
+             }
+        return new ResponseEntity<>("ID not found : "+question.getId(),HttpStatus.NOT_FOUND);
+
+    }
+
+    public ResponseEntity<String> deleteQuestion(int id) {
+        Optional<Question> question=questionDAO.findById(id);
+        if(question.isPresent()){
+            questionDAO.deleteById(id);
+            return new ResponseEntity<>("Question is deleted which ID is : "+id,HttpStatus.OK);
+        }
+        return new ResponseEntity<>("ID not found : "+id,HttpStatus.NOT_FOUND);
     }
 }
